+++ crates/mongodb-agent-common/src/query/execute_query_request.rs
@@ -1,17 +1,16 @@
 use futures::Stream;
 use futures_util::TryStreamExt as _;
-use mongodb::bson;
-use mongodb_support::aggregate::Pipeline;
+use mongodb::{bson, options::AggregateOptions};
+use mongodb_support::aggregate::AggregateCommand;
 use ndc_models::{QueryRequest, QueryResponse};
 use ndc_query_plan::plan_for_query_request;
 use tracing::{instrument, Instrument};
 
-use super::{pipeline::pipeline_for_query_request, response::serialize_query_response};
+use super::{pipeline::command_for_query_request, response::serialize_query_response};
 use crate::{
     interface_types::MongoAgentError,
     mongo_query_plan::{MongoConfiguration, QueryPlan},
     mongodb::{CollectionTrait as _, DatabaseTrait},
-    query::QueryTarget,
 };
 
 type Result<T> = std::result::Result<T, MongoAgentError>;
@@ -31,8 +30,8 @@ pub async fn execute_query_request(
     );
     let query_plan = preprocess_query_request(config, query_request)?;
     tracing::debug!(?query_plan, "abstract query plan");
-    let pipeline = pipeline_for_query_request(config, &query_plan)?;
-    let documents = execute_query_pipeline(database, config, &query_plan, pipeline).await?;
+    let command = command_for_query_request(config, &query_plan)?;
+    let documents = execute_query_command(database, command).await?;
     let response = serialize_query_response(config.extended_json_mode(), &query_plan, documents)?;
     Ok(response)
 }
@@ -47,18 +46,23 @@ fn preprocess_query_request(
 }
 
 #[instrument(name = "Execute Query Pipeline", skip_all, fields(internal.visibility = "user"))]
-async fn execute_query_pipeline(
+async fn execute_query_command(
     database: impl DatabaseTrait,
-    config: &MongoConfiguration,
-    query_plan: &QueryPlan,
-    pipeline: Pipeline,
+    AggregateCommand {
+        collection,
+        pipeline,
+        let_vars,
+    }: AggregateCommand,
 ) -> Result<Vec<bson::Document>> {
-    let target = QueryTarget::for_request(config, query_plan);
     tracing::debug!(
-        ?target,
+        ?collection,
         pipeline = %serde_json::to_string(&pipeline).unwrap(),
+        let_vars = %serde_json::to_string(&let_vars).unwrap(),
         "executing query"
     );
+
+    let aggregate_options =
+        let_vars.map(|let_vars| AggregateOptions::builder().let_vars(let_vars).build());
 
     // The target of a query request might be a collection, or it might be a native query. In the
     // latter case there is no collection to perform the aggregation against. So instead of sending
@@ -67,12 +71,12 @@ async fn execute_query_pipeline(
     // If the query request includes variable sets then instead of specifying the target collection
     // up front that is deferred until the `$lookup` stage of the aggregation pipeline. That is
     // another case where we call `db.aggregate` instead of `db.<collection>.aggregate`.
-    let documents = match (target.input_collection(), query_plan.has_variables()) {
-        (Some(collection_name), false) => {
+    let documents = match collection {
+        Some(collection_name) => {
             let collection = database.collection(collection_name.as_str());
             collect_response_documents(
                 collection
-                    .aggregate(pipeline, None)
+                    .aggregate(pipeline, aggregate_options)
                     .instrument(tracing::info_span!(
                         "MongoDB Aggregate Command",
                         internal.visibility = "user"
@@ -81,10 +85,10 @@ async fn execute_query_pipeline(
             )
             .await
         }
-        _ => {
+        None => {
             collect_response_documents(
                 database
-                    .aggregate(pipeline, None)
+                    .aggregate(pipeline, aggregate_options)
                     .instrument(tracing::info_span!(
                         "MongoDB Aggregate Command",
                         internal.visibility = "user"